+++ contracts/interfaces/mcd/IJug.sol
@@ -5,10 +5,10 @@ pragma solidity =0.8.24;
 abstract contract IJug {
     struct Ilk {
         uint256 duty;
-        uint256  rho;
+        uint256 rho;
     }
 
-    mapping (bytes32 => Ilk) public ilks;
+    mapping(bytes32 => Ilk) public ilks;
 
-    function drip(bytes32) public virtual returns (uint);
+    function drip(bytes32) public virtual returns (uint256);
 }
